improve build deploy to optinally clean and not always stop docker

+++ WeddingWebsite/build_deploy.sh
@@ -1,6 +1,19 @@
 #!/usr/bin/env bash
 
-sudo systemctl start docker
+# Argument count check
+if [ "$#" -gt 3 ]; then
+	echo "Error: Too many arguments. Usage: $0 [start] [keepon] [clean]"
+	exit 1
+fi
+
+# Check for 'start' in any argument
+if [[ "$1" == "start" || "$2" == "start" || "$3" == "start" ]]; then
+	sudo systemctl start docker
+fi
+
+if [[ "$1" == "clean" || "$2" == "clean" || "$3" == "clean" ]]; then
+    sudo docker rmi -f $(sudo docker images --filter "dangling=true" -q --no-trunc)
+fi
 
 sudo docker build -t weddingwebsite .
 
@@ -8,4 +21,9 @@ sudo docker tag weddingwebsite registry.digitalocean.com/octopus-containers/wedd
 
 sudo docker push registry.digitalocean.com/octopus-containers/weddingwebsite
 
-sudo systemctl stop docker
+# Check for 'keepon' in any argument
+if [[ "$1" == "keepon" || "$2" == "keepon" || "$3" == "keepon" ]]; then
+    :  
+else
+    sudo systemctl stop docker
+fi
